requirements.txt
aiogram
biopython